+++ aws/session/custom_ca_bundle_test.go
@@ -6,7 +6,6 @@ import (
 	"net"
 	"net/http"
 	"os"
-	"runtime"
 	"strings"
 	"testing"
 	"time"
@@ -41,17 +40,7 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-// Skips the unit test on travis environment.
-func skipTravisTest(t *testing.T) {
-	// Travis windows environment fails with TLS error when trying to run the unit test.
-	if ver, goos := os.Getenv("TRAVIS_GO_VERSION"), runtime.GOOS; len(ver) != 0 && goos == "windows" {
-		t.Skipf("skipping test, not functional with %s, %s", ver, goos)
-	}
-}
-
 func TestNewSession_WithCustomCABundle_Env(t *testing.T) {
-	skipTravisTest(t)
-
 	restoreEnvFn := initSessionTestEnv()
 	defer restoreEnvFn()
 
@@ -104,8 +93,6 @@ func TestNewSession_WithCustomCABundle_EnvNotExists(t *testing.T) {
 }
 
 func TestNewSession_WithCustomCABundle_Option(t *testing.T) {
-	skipTravisTest(t)
-
 	restoreEnvFn := initSessionTestEnv()
 	defer restoreEnvFn()
 
@@ -141,8 +128,6 @@ func TestNewSession_WithCustomCABundle_Option(t *testing.T) {
 }
 
 func TestNewSession_WithCustomCABundle_HTTPProxyAvailable(t *testing.T) {
-	skipTravisTest(t)
-
 	restoreEnvFn := initSessionTestEnv()
 	defer restoreEnvFn()
 
@@ -171,8 +156,6 @@ func TestNewSession_WithCustomCABundle_HTTPProxyAvailable(t *testing.T) {
 }
 
 func TestNewSession_WithCustomCABundle_OptionPriority(t *testing.T) {
-	skipTravisTest(t)
-
 	restoreEnvFn := initSessionTestEnv()
 	defer restoreEnvFn()
 
@@ -243,8 +226,6 @@ func TestNewSession_WithCustomCABundle_UnsupportedTransport(t *testing.T) {
 }
 
 func TestNewSession_WithCustomCABundle_TransportSet(t *testing.T) {
-	skipTravisTest(t)
-
 	restoreEnvFn := initSessionTestEnv()
 	defer restoreEnvFn()