+++ awstesting/unit/unit.go
@@ -2,12 +2,16 @@
 package unit
 
 import (
+	"time"
+
 	"github.com/IBM/ibm-cos-sdk-go/aws"
 	"github.com/IBM/ibm-cos-sdk-go/aws/credentials"
 	"github.com/IBM/ibm-cos-sdk-go/aws/session"
 )
 
 // Session is a shared session for unit tests to use.
-var Session = session.Must(session.NewSession(aws.NewConfig().
-	WithCredentials(credentials.NewStaticCredentials("AKID", "SECRET", "SESSION")).
-	WithRegion("mock-region")))
+var Session = session.Must(session.NewSession(&aws.Config{
+	Credentials: credentials.NewStaticCredentials("AKID", "SECRET", "SESSION"),
+	Region:      aws.String("mock-region"),
+	SleepDelay:  func(time.Duration) {},
+}))