+++ private/model/api/shape_alias.go
@@ -1,19 +1,19 @@
 package api
 
 var shapeNameAliases = map[string]map[string]string{
-	"APIGateway": {
+	"APIGateway": map[string]string{
 		"RequestValidator": "UpdateRequestValidatorOutput",
 		"VpcLink":          "UpdateVpcLinkOutput",
 		"GatewayResponse":  "UpdateGatewayResponseOutput",
 	},
-	"Lambda": {
+	"Lambda": map[string]string{
 		"Concurrency": "PutFunctionConcurrencyOutput",
 	},
-	"Neptune": {
+	"Neptune": map[string]string{
 		"DBClusterParameterGroupNameMessage": "ResetDBClusterParameterGroupOutput",
 		"DBParameterGroupNameMessage":        "ResetDBParameterGroupOutput",
 	},
-	"RDS": {
+	"RDS": map[string]string{
 		"DBClusterBacktrack": "BacktrackDBClusterOutput",
 	},
 }